add property basic (code need cleanup)

+++ app/src/main/java/com/odauday/data/remote/property/model/GeoLocation.java
@@ -45,8 +45,8 @@ public class GeoLocation implements Parcelable {
     }
     
     protected GeoLocation(Parcel in) {
-        latitude = in.readFloat();
-        longitude = in.readFloat();
+        latitude = in.readDouble();
+        longitude = in.readDouble();
         name = in.readString();
     }
